Quick update... well, a lot of updates
So, I mostly made this for myself and a few friends, and kinda forgot that I posted this to GitHub to begin with. Then I realized how out of date the code is and isn't even fuctional with the current version of discord.py. This should now be fuctional and quite a few new features. If you require assistance, I'll consider putting together a readme file in the future detailing what needs to be subbed out for what.